Moved HAR tests into HarTest class. Added additional HAR tests. Fixed minor issue with incorrect HAR pageCount.
